flask_app/requirements.txt
discord.py
python-dotenv
flask
requests
mysql-connector-python
apscheduler